# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminSessionResponse": {
        "description": "Opaque session material consumed only by the BFF server.",
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "rotated": {
            "title": "Rotated",
            "type": "boolean"
          },
          "session_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Token"
          }
        },
        "required": [
          "expires_at",
          "csrf_token",
          "rotated"
        ],
        "title": "AdminSessionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginPayload": {
        "description": "Administrator login JSON body sent only by the BFF.",
        "properties": {
          "client_ip": {
            "minLength": 1,
            "title": "Client Ip",
            "type": "string"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "password",
          "client_ip"
        ],
        "title": "LoginPayload",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Prediction-market community monitor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "description": "Verify an administrator password and create an opaque session.",
        "operationId": "login_v1_auth_login_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Revoke the current session after same-origin CSRF validation.",
        "operationId": "logout_v1_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Session",
            "required": true,
            "schema": {
              "title": "X-Admin-Session",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "title": "X-Csrf-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "header",
            "name": "referer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Referer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/session": {
      "get": {
        "description": "Validate and rotate an opaque administrator session when required.",
        "operationId": "session_v1_auth_session_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Admin-Session",
            "required": true,
            "schema": {
              "title": "X-Admin-Session",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}